// src/models/articolo.ts
export interface Articolo {
  nome: string
  descrizione: string
  immagine: string
  prezzo: number
  quantita: number

  // constructor (
  //   nome: string,
  //   descrizione: string,
  //   immagine: string,
  //   prezzo: number,
  //   quantita: number
  // ) {
  //   this.nome = nome
  //   this.descrizione = descrizione
  //   this.immagine = immagine
  //   this.prezzo = prezzo
  //   this.quantita = quantita
  // }
}